Move Serializable to base
